examples/tilt: split campaign in two, removed obsolete file
